Verify PIV PIN before we cache it; Fix lint for pincache.go with build tag.
